requirements.txt
pypdf
langchain
langchain-community
langchain-google-genai>=2.0.0
langchain-text-splitters
streamlit
chromadb
google-generativeai
python-dotenv